+++ Changelog.md
@@ -1,6 +1,11 @@
 ### 2.99.0.rc1 Development
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.99.0.beta2...2-99-maintenance)
 
+Deprecations:
+
+* Deprecate `RSpec::Core::Pending::PendingDeclaredInExample`, use
+  `SkipDeclaredInExample` instead. (Xavier Shay)
+
 ### 2.99.0.beta2 / 2014-02-17
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.99.0.beta1...v2.99.0.beta2)